+++ frontend/lost/src/containers/DataSources/DSTable.jsx
@@ -11,7 +11,7 @@ import LostFileBrowser from '../../components/FileBrowser/LostFileBrowser'
 import * as Notification from '../../components/Notification'
 import EditDSModal from './EditDSModal'
 import { CBadge } from '@coreui/react'
-import * as fbAPI from '../../api/fb'
+import * as fbAPI from '../../api/file_browser'
 import BaseContainer from '../../components/BaseContainer'
 import CoreDataTable from '../../components/CoreDataTable'
 import CoreIconButton from '../../components/CoreIconButton'
